@@ -1,8 +1,9 @@
 import json
-from tests.basetest import Basetest
+from pathlib import Path
 
 from slides.slidewalker import SlideWalker
-from pathlib import Path
+from tests.basetest import Basetest
+
 
 class TestSlideWalker(Basetest):
     """
@@ -15,28 +16,28 @@ class TestSlideWalker(Basetest):
         """
         Basetest.setUp(self, debug=debug, profile=profile)
         self.debug = debug
-        base_path=Path(__file__).parent.parent
-        self.slidedir=f"{base_path}/examples/semanticslides"
+        base_path = Path(__file__).parent.parent
+        self.slidedir = f"{base_path}/examples/semanticslides"
         pass
-   
+
     def test_slidewalker(self):
         """
         simple test for slidewalker
         """
-        debug=self.debug
-        #debug=True
+        debug = self.debug
+        # debug=True
         slidewalker = SlideWalker(self.slidedir)
-        json_str=slidewalker.dumpInfoToString("json", excludeHiddenSlides=True)
+        json_str = slidewalker.dumpInfoToString("json", excludeHiddenSlides=True)
         if debug:
             print(json_str)
-        pres_dict=json.loads(json_str)
-        ppt_file="SemanticSlides.pptx"
+        pres_dict = json.loads(json_str)
+        ppt_file = "SemanticSlides.pptx"
         self.assertTrue(ppt_file in pres_dict)
-        pres=pres_dict[ppt_file]
+        pres = pres_dict[ppt_file]
         self.assertTrue("slides" in pres)
-        slides=pres["slides"]
-        self.assertTrue(len(slides)>1)
+        slides = pres["slides"]
+        self.assertTrue(len(slides) > 1)
         for slide in slides:
-            for attr in ["page","pdf_page","title","name","text","notes"]:
+            for attr in ["page", "pdf_page", "title", "name", "text", "notes"]:
                 self.assertTrue(attr in slide)
         pass
